requirements.txt
Django==5.1.3
django-ninja==1.3.0
rav==0.0.9
python-decouple==3.8
yfinance==0.2.50
